# toeic-learning/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Toeic
# during dev, should point to your local machine
host: 103.114.107.16:8003
...
basePath: /api/learning
...
securityDefinitions:
  auth0:
    type: apiKey
    name: Authorization
    in: header
    description: "Authorization logged to application"
schemes:
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /lessons:
    x-swagger-router-controller: lesson
    get:
      description: Returns all lesson
      operationId: getLessons
      tags:
        - lesson
      security:
        - auth0: []
      parameters:
        - in: query
          name: limit
          type: integer
          description: limit value return
        - in: query
          name: page
          type: integer
          description: page get
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MessageResp"
        default:
          description: Error5
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Add new lesson
      operationId: addLesson
      tags:
        - lesson
      security:
        - auth0: []
      parameters:
        - in: body
          name: body
          description: parameters with content, topic, category, title in new lesson .
          schema:
            $ref: '#/definitions/LessonReq'
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MessageResp"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /lessons/{lessonId}:
    x-swagger-router-controller: lesson
    get:
      description: Returns lesson info with parameters lessonId
      operationId: findOneLesson
      tags:
        - lesson
      security:
        - auth0: []
      parameters:
        - in: path
          name: lessonId
          type: string
          required: true
          description: id lesson
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MessageResp"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Update lesson info
      operationId: updateLesson
      tags:
        - lesson
      security:
        - auth0: []
      parameters:
        - in: path
          name: lessonId
          type: string
          required: true
          description: id lesson
        - in: body
          name: body
          description: parameters with content, topic, category, title in new lesson.
          schema:
            $ref: '#/definitions/LessonReq'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MessageResp"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: delete lesson with id
      operationId: deleteLesson
      tags:
        - lesson
      security:
        - auth0: []
      parameters:
        - in: path
          name: lessonId
          type: string
          required: true
          description: id lesson
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MessageResp"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  LessonReq:
    required:
      - title
      - content
      - topicId
      - categoryId
    properties:
      content:
        type: string
      topicId:
        type: string
      categoryId:
        type: string
      title:
        type: string
...
  MessageResp:
    type: object
    required: 
      - success
    properties:
      success:
        type: boolean
      message:
        type: string
      value:
        type: object
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string